@@ -11,26 +11,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
+    "@angular/animations": "^19.1.2",
+    "@angular/common": "^19.1.2",
+    "@angular/compiler": "^19.1.2",
+    "@angular/core": "^19.1.2",
     "@angular/fire": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/platform-server": "^19.0.0",
-    "@angular/router": "^19.0.0",
-    "@angular/ssr": "^19.0.6",
+    "@angular/forms": "^19.1.2",
+    "@angular/platform-browser": "^19.1.2",
+    "@angular/platform-browser-dynamic": "^19.1.2",
+    "@angular/platform-server": "^19.1.2",
+    "@angular/router": "^19.1.2",
+    "@angular/ssr": "^19.1.3",
     "express": "^4.18.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.6",
-    "@angular/cli": "^19.0.6",
-    "@angular/compiler-cli": "^19.0.0",
+    "@angular-devkit/build-angular": "^19.1.3",
+    "@angular/cli": "^19.1.3",
+    "@angular/compiler-cli": "^19.1.2",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
